Convert datetime row generators to hooks
